chore: move utils file outside of src directory - Moved util file to the root directory for better organization and accessibility. - Created testUtils.js for testing purposes.
