+++ src/views/OutputView.js
@@ -10,7 +10,6 @@ class OutputView {
     cars.forEach((car) => {
       Console.print(`${car.getName()} : ${'-'.repeat(car.getPosition())}`);
     });
-    Console.print('');
   }
 
   printWinners(winners) {
